x/vnpay/requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.1
python-dotenv==1.0.1
setuptools==75.6.0
